--- src/containers/CreateRecordContainer/index.tsx
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from './CreateRecordContainer';
