+++ src/containers/UpsertRecordContainer/index.tsx
@@ -1,0 +1,1 @@
+export { default, getServerSideProps } from './UpsertRecordContainer';
